refector: 프로필 페이지 분리
